Struct crowbook::book::Book

pub struct Book {
    pub chapters: Vec<(Number, Vec<Token>)>,
    // some fields omitted
}

Fields

chapters

Methods

impl Book

fn description(md: bool) -> String

Returns a description for all options

fn new() -> Book

Creates a new Book with default options

fn println(&self, s: &str)

Prints to stderr

fn debug(&self, s: &str)

Prints to stderr but only if verbose is set to true

fn new_from_file(filename: &str, verbose: bool) -> Result<Book>

Creates a new book from a file

This method also changes the current directory to the one of this file

fn get_mapbuilder(&self, format: &str) -> MapBuilder

Returns a MapBuilder, to be used (and completed) for templating

fn clean(&self, text: String) -> String

Either clean a string or does nothing

fn get_cleaner(&self) -> Option<Box<Cleaner>>

Return a Box corresponding to the appropriate cleaning method, or None

fn get_header(&self, n: i32, title: &str) -> Result<String>

Returns the string corresponding to a number, title, and the numbering template

fn get_option(&self, key: &str) -> Result<&BookOption>

get an option

fn get_str(&self, key: &str) -> Result<&str>

gets a string option as str

fn get_bool(&self, key: &str) -> Result<bool>

gets a bool option

fn get_char(&self, key: &str) -> Result<char>

gets a char option

fn get_i32(&self, key: &str) -> Result<i32>

gets an int option

fn set_option(&mut self, key: &str, value: &str) -> Result<()>

Sets an option

fn set_from_config(&mut self, s: &str) -> Result<()>

Sets options and load chapters according to configuration file

A line with "option: value" sets the option to value + chapter_name.md adds the (default numbered) chapter - chapter_name.md adds the (unnumbered) chapter 3. chapter_name.md adds the (custom numbered) chapter

fn render_pdf(&self) -> Result<()>

Render book to pdf according to book options

fn render_epub(&self) -> Result<()>

Render book to epub according to book options

fn render_odt(&self) -> Result<()>

Render book to odt according to book options

fn render_html<T: Write>(&self, f: &mut T) -> Result<()>

Render book to html according to book options

fn render_tex<T: Write>(&self, f: &mut T) -> Result<()>

Render book to pdf according to book options

fn render_all(&self) -> Result<()>

Generates output files acccording to book options

fn add_chapter(&mut self, number: Number, file: &str) -> Result<()>

File: location of the file for this chapter

fn get_template(&self, template: &str) -> Result<Cow<'static, str>>

Returns the template (default or modified version)