Enum crowbook::Token

pub enum Token {
    Str(String),
    Paragraph(Vec<Token>),
    Header(i32, Vec<Token>),
    Emphasis(Vec<Token>),
    Strong(Vec<Token>),
    Code(Vec<Token>),
    BlockQuote(Vec<Token>),
    CodeBlock(String, Vec<Token>),
    List(Vec<Token>),
    OrderedList(usize, Vec<Token>),
    Item(Vec<Token>),
    Table(i32, Vec<Token>),
    TableHead(Vec<Token>),
    TableRow(Vec<Token>),
    TableCell(Vec<Token>),
    Footnote(Vec<Token>),
    Rule,
    SoftBreak,
    HardBreak,
    Link(String, String, Vec<Token>),
    Image(String, String, Vec<Token>),
}

A single token representing a Markdown element.

A Markdown document is, thus, a Vec of Tokens.

Variants

Str

The most simple element, containing a String

Paragraph

A paragraph, containing a list of elements

Header

A header with a header number and the title

Emphasis

Emphasis, a.k.a. italics

Strong

Strong, a.k.a. bold

Code

Code, a.k.a. verbatim

BlockQuote

A quote

CodeBlock

Code block with language and content

List

Unordered list, with a vector of Items

OrderedList

Ordered list, with a starting number, and a list of Items

Item

Item of a list

Table

Table with number of rows, and a list of TableHead and TableRows

TableHead

Table header, contains TableCells

TableRow

Row of a table, contains TableCells

TableCell

Cell of a table

Footnote

A footnote, contains the content it is pointing to.

Rule

Horizontal rule

SoftBreak

Softbreak, usually rendered by a space

HardBreak

Hardbreak

Link

A link with an url, a title, and the linked text

Image

An image with a source url, a title and an alt tex

Methods

impl Token

fn inner(&self) -> Option<&[Token]>

Returns the inner list of tokens contained in this token (if any)

fn inner_mut(&mut self) -> Option<&mut Vec<Token>>

Returns the inner list of tokens contained in this token (if any) (mutable version)