Struct crowbook::Book

pub struct Book {
    pub chapters: Vec<(Number, Vec<Token>)>,
    pub filenames: Vec<String>,
    pub options: BookOptions,
    pub root: PathBuf,
}

A Book.

Probably the central structure for of Crowbook, as it is the one that calls the other ones.

It has the tasks of loading a configuration file, loading chapters and using Parserto parse them, and then calling various renderers (HtmlRendrer, LatexRenderer, EpubRenderer and/or OdtRenderer) to convert the AST into documents.

Examples

use crowbook::{Book, Number};
// Create an empty book
let mut book = Book::new();
// Set some options
book.options.set("author", "Joan Doe");
book.options.set("title", "An untitled book");
book.options.set("lang", "en");

// Add content to the book
book.add_chapter_as_str(Number::Default, "# The beginning#\nBla, bla, bla").unwrap();

// Render the book as html to stdout
book.render_html(&mut std::io::stdout()).unwrap();

Fields

chapters

Internal structure. You should not accesss this directly except if you are writing a new renderer.

filenames

A list of the filenames of the chapters

options

Options of the book

root

Root path of the book

Methods

impl Book

fn new() -> Book

Creates a new, empty Book with default options

fn new_from_file(filename: &str, verbose: bool) -> Result<Book>

Creates a new book from a file

Arguments

filename: the path of file to load. The directory of this file is used as a "root" directory for all paths referenced in books, whether chapter files, templates, cover images, and so on.

verbose: sets the book to verbose mode even if the file's doesn't specify it or specifies verbose: false.

fn set_from_config(&mut self, s: &str) -> Result<()>

Sets options and load chapters according to configuration file

A line with "option: value" sets the option to value + chapter_name.md adds the (default numbered) chapter - chapter_name.md adds the (unnumbered) chapter 3. chapter_name.md adds the (custom numbered) chapter

fn render_all(&self) -> Result<()>

Generates output files acccording to book options

fn render_pdf(&self) -> Result<()>

Render book to pdf according to book options

fn render_epub(&self) -> Result<()>

Render book to epub according to book options

fn render_odt(&self) -> Result<()>

Render book to odt according to book options

fn render_html<T: Write>(&self, f: &mut T) -> Result<()>

Render book to html according to book options

fn render_tex<T: Write>(&self, f: &mut T) -> Result<()>

Render book to pdf according to book options

fn add_chapter(&mut self, number: Number, file: &str) -> Result<()>

Adds a chapter, as a file name, to the book

Book will then parse the file and store the AST (i.e., a vector of Tokens).

Arguments

number: specifies if the chapter must be numbered, not numbered, or if its title must be hidden. See Number.

file: path of the file for this chapter

Returns an error if file does not exist, could not be read, of if there was some error parsing it.

fn add_chapter_as_str(&mut self, number: Number, content: &str) -> Result<()>

Adds a chapter, as a string, to the book

Book will then parse the string and store the AST (i.e., a vector of Tokens).

Arguments

number: specifies if the chapter must be numbered, not numbered, or if its title must be hidden. See Number.

content: the content of the chapter.

Returns an error if there was some errror parsing content.

fn clean(&self, text: String) -> String

Either clean a string or does nothing, according to book lang and autoclean options

fn get_template(&self, template: &str) -> Result<Cow<'static, str>>

Returns the template (default or modified version)

fn get_header(&self, n: i32, title: &str) -> Result<String>

Returns the string corresponding to a number, title, and the numbering template

fn println(&self, s: &str)

Prints a message to stderr

fn debug(&self, s: &str)

Prints a message to stderr if verbose is set to true

fn get_mapbuilder(&self, format: &str) -> MapBuilder

Returns a MapBuilder (used by Mustache for templating), to be used (and completed) by renderers. It fills it with the followings strings, corresponding to the matching Book options:

"author"

"title"

"lang"