Struct crowbook::Parser

pub struct Parser {
    // some fields omitted
}

A parser that reads markdown and convert it to AST (a vector of Tokens)

This AST can then be used by various renderes.

As this Parser uses Pulldown-cmark's one, it should be able to parse most valid CommonMark variant of Markdown.

Compared to other Markdown parser, it might fail more often on invalid code, e.g. footnotes references that are not defined anywhere.

Examples

use crowbook::Parser;
let mut parser = Parser::new();
let result = parser.parse("Some *valid* Markdown[^1]\n\n[^1]: with a valid footnote");
assert!(result.is_ok());

use crowbook::Parser;
let mut parser = Parser::new();
let result = parser.parse("Some footnote pointing to nothing[^1] ");
assert!(result.is_err());

Methods

impl Parser

fn new() -> Parser

Creates a parser

fn parse_file<P: AsRef<Path>>(&mut self, filename: P) -> Result<Vec<Token>>

Parse a file and returns an AST or an error

fn parse(&mut self, s: &str) -> Result<Vec<Token>>

Parse a string and returns an AST, or an Error.