Struct crowbook::book::Book

pub struct Book {
    pub lang: String,
    pub author: String,
    pub title: String,
    pub description: Option<String>,
    pub subject: Option<String>,
    pub cover: Option<String>,
    pub output_epub: Option<String>,
    pub output_html: Option<String>,
    pub output_pdf: Option<String>,
    pub output_tex: Option<String>,
    pub output_odt: Option<String>,
    pub temp_dir: String,
    pub chapters: Vec<(Number, Vec<Token>)>,
    pub numbering: bool,
    pub autoclean: bool,
    pub nb_char: char,
    pub numbering_template: String,
    pub verbose: bool,
    pub tex_command: String,
    pub tex_links_as_footnotes: bool,
    pub tex_template: Option<String>,
    pub epub_css: Option<String>,
    pub epub_template: Option<String>,
    pub epub_version: u8,
    pub html_template: Option<String>,
    pub html_css: Option<String>,
}

Fields

lang

author

title

description

subject

cover

output_epub

output_html

output_pdf

output_tex

output_odt

temp_dir

chapters

numbering

autoclean

nb_char

numbering_template

verbose

tex_command

tex_links_as_footnotes

tex_template

epub_css

epub_template

epub_version

html_template

html_css

Methods

impl Book

fn new() -> Book

fn new_from_file(filename: &str) -> Result<Book>

Creates a new book from a file

This method also changes the current directory to the one of this file

fn get_mapbuilder(&self, format: &str) -> MapBuilder

Returns a MapBuilder, to be used (and completed) for templating

fn get_cleaner(&self) -> Option<Box<Cleaner>>

Return a Box corresponding to the appropriate cleaning method, or None

fn get_header(&self, n: i32, title: &str) -> Result<String>

Returns the string corresponding to a number, title, and the numbering template

fn set_from_config(&mut self, s: &str) -> Result<()>

Sets options according to configuration file

A line with "option: value" sets the option to value + chapter_name.md adds the (default numbered) chapter - chapter_name.md adds the (unnumbered) chapter 3. chapter_name.md adds the (custom numbered) chapter

fn render_pdf(&self, file: &str) -> Result<()>

Render book to pdf according to book options

fn render_epub(&self) -> Result<()>

Render book to epub according to book options

fn render_odt(&self) -> Result<()>

Render book to odt according to book options

fn render_html(&self, file: &str) -> Result<()>

Render book to html according to book options

fn render_tex(&self, file: &str) -> Result<()>

Render book to pdf according to book options

fn render_all(&self) -> Result<()>

Generates output files acccording to book options

fn add_chapter(&mut self, number: Number, file: &str) -> Result<()>

File: location of the file for this chapter

fn get_template(&self, template: &str) -> Result<Cow<'static, str>>

Returns the template (default or modified version)